Fix player menu track selection
Because the menus use setGroupCheckable with exclusive=true, it isn't necessary to uncheck the other items.
Furthermore, while transcoding, the media source is updated and the menu rebuilt (with the new selections already). This causes the clickedItem to become outdated/invalid in the fragment.onXTrackSelected callback. As a result, the selection will be invalidated.
To fix this, we assume that selections are always successful and immediately apply the selection to the clicked item before notifying the fragment.